Allow any name for the base file in the web space
The whole app runs from a python file that uses the foodlog package. The
HTML pages created need to link back that original file. It wsa hard
coded in a number of them that the original file name was run.py.
This change takes out that hard coding, and makes the links based on the
environment variable set in cgi-bin to indicate the script file name
being run.