Copying picture when copying course
If you bring up a dish in the edit page, from search or off a report
it can have an image in it. Now if you use the copy button, the
image id will be copied into the dish, so it defaults that way
when you save. If you upload and image, however, that image will be
used instead of the one from the copy